+++ bin/generic-install.sh
@@ -139,9 +139,9 @@ chmod +x ${PIP3}
 echo "Downloading Pip packages to $LIB_DIR..."
 ${PIP3} download -d $LIB_DIR --no-deps little_brother==0.4.25
 
-${PIP3} download -d $LIB_DIR --no-deps python_base_app==0.2.42
+${PIP3} download -d $LIB_DIR --no-deps python_base_app==0.2.43
 
-${PIP3} download -d $LIB_DIR --no-deps some_flask_helpers==0.2.2
+${PIP3} download -d $LIB_DIR --no-deps some_flask_helpers==0.2.3
 
 
 echo "Checking if all Pip packages have been downloaded to $LIB_DIR..."
@@ -153,20 +153,20 @@ else
   echo "Package little-brother-0.4.25.tar.gz was found."
 fi
 
-if [ ! -f $LIB_DIR/python-base-app-0.2.42.tar.gz ] ; then
-  echo "ERROR: package python-base-app-0.2.42.tar.gz not found in $LIB_DIR!"
+if [ ! -f $LIB_DIR/python-base-app-0.2.43.tar.gz ] ; then
+  echo "ERROR: package python-base-app-0.2.43.tar.gz not found in $LIB_DIR!"
   echo "Download from test.pypi.org and execute again."
   exit 2
 else
-  echo "Package python-base-app-0.2.42.tar.gz was found."
+  echo "Package python-base-app-0.2.43.tar.gz was found."
 fi
 
-if [ ! -f $LIB_DIR/some-flask-helpers-0.2.2.tar.gz ] ; then
-  echo "ERROR: package some-flask-helpers-0.2.2.tar.gz not found in $LIB_DIR!"
+if [ ! -f $LIB_DIR/some-flask-helpers-0.2.3.tar.gz ] ; then
+  echo "ERROR: package some-flask-helpers-0.2.3.tar.gz not found in $LIB_DIR!"
   echo "Download from test.pypi.org and execute again."
   exit 2
 else
-  echo "Package some-flask-helpers-0.2.2.tar.gz was found."
+  echo "Package some-flask-helpers-0.2.3.tar.gz was found."
 fi
 
 if [ "$RUNNING_IN_DOCKER" == "" ] ; then
@@ -277,21 +277,21 @@ chmod og-rwx /etc/little-brother/little-brother.config
 ${PIP3} install wheel # setuptools
 echo "Installing PIP packages..."
 echo "  * little-brother-0.4.25.tar.gz"
-echo "  * python-base-app-0.2.42.tar.gz"
-echo "  * some-flask-helpers-0.2.2.tar.gz"
+echo "  * python-base-app-0.2.43.tar.gz"
+echo "  * some-flask-helpers-0.2.3.tar.gz"
 # see https://stackoverflow.com/questions/19548957/can-i-force-pip-to-reinstall-the-current-version
 ${PIP3} install --upgrade \
      ${LIB_DIR}/little-brother-0.4.25.tar.gz\
-     ${LIB_DIR}/python-base-app-0.2.42.tar.gz\
-     ${LIB_DIR}/some-flask-helpers-0.2.2.tar.gz
+     ${LIB_DIR}/python-base-app-0.2.43.tar.gz\
+     ${LIB_DIR}/some-flask-helpers-0.2.3.tar.gz
 
 
 echo "Removing installation file ${LIB_DIR}/little-brother-0.4.25.tar.gz..."
 rm ${LIB_DIR}/little-brother-0.4.25.tar.gz
-echo "Removing installation file ${LIB_DIR}/python-base-app-0.2.42.tar.gz..."
-rm ${LIB_DIR}/python-base-app-0.2.42.tar.gz
-echo "Removing installation file ${LIB_DIR}/some-flask-helpers-0.2.2.tar.gz..."
-rm ${LIB_DIR}/some-flask-helpers-0.2.2.tar.gz
+echo "Removing installation file ${LIB_DIR}/python-base-app-0.2.43.tar.gz..."
+rm ${LIB_DIR}/python-base-app-0.2.43.tar.gz
+echo "Removing installation file ${LIB_DIR}/some-flask-helpers-0.2.3.tar.gz..."
+rm ${LIB_DIR}/some-flask-helpers-0.2.3.tar.gz
 if [ "$RUNNING_IN_DOCKER" == "" ] ; then
   echo "Execute systemctl daemon-reload..."
   set +e