+++ CHANGES.md
@@ -7,6 +7,8 @@ This document lists all changes of `LittleBrother` with the most recent changes 
 ## Version 0.4.25 Revision 131 (July 3rd, 2022)
 
 * Re-activate all images (left deactivated by mistake for testing purposes)
+* Upgrade to package `Flask` > 2
+* Upgrade to `some-flask-helpers` 0.2.3
 
 ## Version 0.4.24 Revision 131 (July 3rd, 2022)